+++ src/lib.rs
@@ -207,9 +207,6 @@ mod testutil {
     impl Drop for RandomFile {
         fn drop(&mut self) {
             let _ = std::fs::remove_file(&self.path);
-            let mut sidecar = self.path.as_os_str().to_owned();
-            sidecar.push(".tagdata");
-            let _ = std::fs::remove_dir_all(std::path::PathBuf::from(sidecar));
         }
     }